requirements.txt
fastapi
uvicorn
insightface
onnxruntime
Pillow
opencv-contrib-python-headless==4.9.0.80
numpy<2.0
